DustChip component and Tool tip for Tx Preview
