+++ apps/coordinator/src/components/ScriptExplorer/UTXOSet.jsx
@@ -14,14 +14,12 @@ import {
   TableCell,
   Typography,
   Checkbox,
-  Chip,
 } from "@mui/material";
 import { OpenInNew } from "@mui/icons-material";
 import BigNumber from "bignumber.js";
 import { externalLink } from "utils/ExternalLink";
 import Copyable from "../Copyable";
-import { WasteMetrics } from "@caravan/health";
-import { getWalletConfig } from "../../selectors/wallet";
+import DustChip from "./DustChip";
 
 // Actions
 import { setInputs as setInputsAction } from "../../actions/transactionActions";
@@ -173,49 +171,13 @@ class UTXOSet extends React.Component {
     }
   };
 
-  getChipProperties(amountSats, feeRate, walletConfig) {
-    const wasteMetrics = new WasteMetrics();
-    const scriptType = walletConfig.addressType;
-    const config = {
-      requiredSignerCount: walletConfig.quorum.requiredSigners,
-      totalSignerCount: walletConfig.quorum.totalSigners,
-    };
-    const { lowerLimit, upperLimit } = wasteMetrics.calculateDustLimits(
-      feeRate,
-      scriptType,
-      config,
-    );
-    if (amountSats <= lowerLimit) {
-      return {
-        color: "error",
-        label: "dust",
-      };
-    } else if (amountSats > lowerLimit && amountSats <= upperLimit) {
-      return {
-        color: "warning",
-        label: "warning",
-      };
-    } else {
-      return {
-        color: "success",
-        label: "economical",
-      };
-    }
-  }
-
   renderInputs = () => {
-    const { network, showSelection, finalizedOutputs, walletConfig, feeRate } =
-      this.props;
+    const { network, showSelection, finalizedOutputs, feeRate } = this.props;
     const { localInputs } = this.state;
     return localInputs.map((input, inputIndex) => {
       const confirmedStyle = `${styles.utxoTxid}${
         input.confirmed ? "" : ` ${styles.unconfirmed}`
       }`;
-      const { color, label } = this.getChipProperties(
-        input.amountSats,
-        feeRate,
-        walletConfig,
-      );
       const confirmedTitle = input.confirmed ? "confirmed" : "unconfirmed";
       return (
         <TableRow hover key={input.txid}>
@@ -243,7 +205,7 @@ class UTXOSet extends React.Component {
             <Copyable text={satoshisToBitcoins(input.amountSats)} />
           </TableCell>
           <TableCell>
-            <Chip color={color} label={label}></Chip>
+            <DustChip amountSats={input.amountSats} feeRate={feeRate} />
           </TableCell>
           <TableCell>
             {externalLink(
@@ -328,7 +290,6 @@ UTXOSet.propTypes = {
   setSpendCheckbox: PropTypes.func,
   autoSpend: PropTypes.bool.isRequired,
   feeRate: PropTypes.string,
-  walletConfig: PropTypes.object,
 };
 
 UTXOSet.defaultProps = {
@@ -348,7 +309,6 @@ function mapStateToProps(state) {
     autoSpend: state.spend.transaction.autoSpend,
     finalizedOutputs: state.spend.transaction.finalizedOutputs,
     existingTransactionInputs: state.spend.transaction.inputs,
-    walletConfig: getWalletConfig(state),
   };
 }
